imports/exports Metric through common/metric
